+++ src/main/java/com/scope/socialboardweb/dto/table/TableEntityDto.java
@@ -1,4 +1,4 @@
-package com.scope.socialboardweb.dto.test;
+package com.scope.socialboardweb.dto.table;
 
 public interface TableEntityDto {
     public String getFieldValueByName(String fieldName);
