+++ src/main/java/com/scope/socialboardweb/dto/table/UserTableEntityDto.java
@@ -1,4 +1,4 @@
-package com.scope.socialboardweb.dto.test;
+package com.scope.socialboardweb.dto.table;
 
 import com.scope.socialboardweb.domain.User;
 import lombok.Getter;
