+++ src/main/java/com/scope/socialboardweb/service/test/AdminService.java
@@ -5,16 +5,12 @@ import com.scope.socialboardweb.domain.Post;
 import com.scope.socialboardweb.domain.User;
 import com.scope.socialboardweb.dto.LoginRequestDto;
 import com.scope.socialboardweb.dto.LoginResponseDto;
-import com.scope.socialboardweb.dto.test.CommentTableEntityDto;
-import com.scope.socialboardweb.dto.test.PostTableEntityDto;
-import com.scope.socialboardweb.dto.test.UserTableEntityDto;
 import com.scope.socialboardweb.repository.CommentRepository;
 import com.scope.socialboardweb.repository.PostRepository;
 import com.scope.socialboardweb.repository.UserRepository;
 import com.scope.socialboardweb.repository.custom.CustomUserRepository;
 import com.scope.socialboardweb.service.UserService;
 import com.scope.socialboardweb.utils.jwt.JwtTokenProvider;
-import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
